% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age}
\alias{age}
\title{Âge}
\format{
## `age`
Un vecteur de longueur 20.
\describe{
  \item{age}{Âge du participant}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
age
}
\description{
Variable discrète ayant 20 participants.
}
\keyword{datasets}
