% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anxiete}
\alias{anxiete}
\title{Anxiété}
\format{
## `anxiete`
Un vecteur de longueur 20.
\describe{
  \item{anxiete}{Mesure à l'aide d'un test de symptômes d'anxiété}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
anxiete
}
\description{
Variable ordinale à 20 points ayant 20 participants.
}
\keyword{datasets}
