% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_t_to_z.R
\name{compare_t_to_z}
\alias{compare_t_to_z}
\title{Fonction simple pour comparer la distribution t et la distribution normale}
\usage{
compare_t_to_z(df = 1)
}
\arguments{
\item{df}{Nombre de degre de liberte.}
}
\value{
Un graphique
}
\description{
Cette fonction permet de visualiser la différence entre les courbes de t et
de z en changeant le nombre de degré de liberté de t. Plus dl est grand, et plus
elle se rapproche d'une distribution z.
}
\examples{
compare_t_to_z(df = 2)
}
