% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{couleur}
\alias{couleur}
\title{Couleur}
\format{
## `couleur`
Un vecteur de longueur 20.
\describe{
  \item{couleur}{Couleur préférée du participant}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
couleur
}
\description{
Variable nominale à 3 catégories ayant 20 participants.
}
\keyword{datasets}
