% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data01}
\alias{data01}
\title{data01}
\format{
## `data01`
Une base de données quelconque.
\describe{
  \item{data01}{data01}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
data01
}
\description{
Une base de données avec trois variables.
}
\keyword{datasets}
