% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression}
\alias{depression}
\title{Dépression}
\format{
## `depression`
Un vecteur de longueur 20.
\describe{
  \item{depression}{Mesure d'un score de vitesse en seconde}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
depression
}
\description{
Variable ordinale à 40 points ayant 20 participants.
}
\keyword{datasets}
