% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presence}
\alias{presence}
\title{Présence}
\format{
## `presence`
Un vecteur de longueur 20.
\describe{
  \item{presence}{Présence ou absence du participant}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
presence
}
\description{
Variable dichotomique ayant 20 participants.
}
\keyword{datasets}
