% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{score02}
\alias{score02}
\title{Score02}
\format{
## `score02`
Un vecteur de longueur 20.
\describe{
  \item{score02}{Mesure d'un score de vitesse en seconde}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
score02
}
\description{
Variable discrète ayant 20 participants.
}
\keyword{datasets}
