% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab500}
\alias{tab500}
\title{Tableau 500}
\format{
## `tab500`
Un tableau de données ayant 500 participants
\describe{
  \item{id}{Numéro d'identification unique du participant}
  \item{age}{Âge du participant}
  \item{sexe}{Sexe du participant}
  \item{presence}{Présence ou absence du participant}
  \item{temps}{Mesure de temps}
  \item{distance}{Mesure de distance}
  \item{anxiete}{Mesure à l'aide d'un test de symptômes d'anxiété}
  \item{depression}{Mesure à l'aide d'un test de symptômes dépressifs}
  \item{couleur}{Couleur préférée du participant}
  \item{score01}{Mesure d'un score à un temps 1}
  \item{score02}{Mesure d'un score à un temps 2}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
tab500
}
\description{
Un tableau de données ayant 500 participants et donc certaines variables sont corrélées.
}
\keyword{datasets}
