% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_anxiete500}
\alias{var_anxiete500}
\title{Anxiété}
\format{
## `var_anxiete500`
Un vecteur de longueur 500.
\describe{
  \item{var_anxiete500}{Score d'anxiété}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_anxiete500
}
\description{
Variable discrète ayant 500 participants.
}
\keyword{datasets}
