% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_couleur}
\alias{var_couleur}
\title{Couleur}
\format{
## `var_couleur`
Un vecteur de longueur 20.
\describe{
  \item{var_couleur}{Couleur préférée du participant}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_couleur
}
\description{
Variable nominale à 3 catégories ayant 20 participants.
}
\keyword{datasets}
