% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_depression}
\alias{var_depression}
\title{Dépression}
\format{
## `var_depression`
Un vecteur de longueur 20.
\describe{
  \item{var_depression}{Mesure à l'aide d'un test de symptômes dépressifs}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_depression
}
\description{
Variable ordinale à 40 points ayant 20 participants.
}
\keyword{datasets}
