% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_distance}
\alias{var_distance}
\title{Distance}
\format{
## `var_temps`
Un vecteur de longueur 20.
\describe{
  \item{var_distance}{Mesure de distance}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_distance
}
\description{
Variable continue ayant 20 participants.
}
\keyword{datasets}
