% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_presence}
\alias{var_presence}
\title{Présence}
\format{
## `var_presence`
Un vecteur de longueur 20.
\describe{
  \item{var_presence}{Présence ou absence du participant}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_presence
}
\description{
Variable dichotomique ayant 20 participants.
}
\keyword{datasets}
