% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_score01}
\alias{var_score01}
\title{Score01}
\format{
## `var_score01`
Un vecteur de longueur 20.
\describe{
  \item{var_score01}{Mesure d'un score à un temps 1}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_score01
}
\description{
Variable discrète ayant 20 participants.
}
\keyword{datasets}
