% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_score02}
\alias{var_score02}
\title{Score02}
\format{
## `var_score02`
Un vecteur de longueur 20.
\describe{
  \item{var_score02}{Mesure d'un score à un temps 2}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_score02
}
\description{
Variable discrète ayant 20 participants.
}
\keyword{datasets}
