% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_pop.R
\name{var_pop}
\alias{var_pop}
\alias{sd_pop}
\title{Compute population variance and standard deviation}
\usage{
var_pop(x)

sd_pop(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric scalar.
}
\description{
Compute population variance and standard deviation
}
\examples{
x <- rnorm(10)
var_pop(x)
sd_pop(x)
}
