% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_sexe}
\alias{var_sexe}
\title{Sexe}
\format{
## `var_sexe`
Un vecteur de longueur 20.
\describe{
  \item{var_sexe}{Sexe du participant}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_sexe
}
\description{
Variable nominale à deux catégories ayant 20 participants.
}
\keyword{datasets}
