% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_temps}
\alias{var_temps}
\title{Temps}
\format{
## `var_temps`
Un vecteur de longueur 20.
\describe{
  \item{var_temps}{Mesure de temps}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_temps
}
\description{
Variable continue ayant 20 participants.
}
\keyword{datasets}
