% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_temps500}
\alias{var_temps500}
\title{Temps}
\format{
## `var_temps500`
Un vecteur de longueur 500.
\describe{
  \item{var_temps500}{Mesure de temps}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
var_temps500
}
\description{
Variable continue ayant 500 participants.
}
\keyword{datasets}
