% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_dist.R
\name{visualize_area}
\alias{visualize_area}
\title{Visualisation des zones de alpha et beta sous la courbe normale}
\usage{
visualize_area(
  mu_0 = 100,
  sigma_0 = 15,
  mu_1 = 107,
  sigma_1 = NULL,
  n = 50,
  alpha = 0.05,
  fill = "alpha"
)
}
\arguments{
\item{mu_0}{La moyenne de la population}

\item{sigma_0}{L'écart type de la population}

\item{mu_1}{La moyenne de l'échantillon}

\item{sigma_1}{L'écart type de l'échantillon, si différente de sigma_0}

\item{n}{La taille d'échantillon}

\item{alpha}{L'erreur de type II, valeur entre 0 et 1}

\item{fill}{Type de coloration sous la courbe, soit "alpha" (défaut) ou "beta"}
}
\value{
Un graphique.
}
\description{
Visualisation des zones de alpha et beta sous la courbe normale
}
\examples{
visualize_area()
}
