% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vitesse}
\alias{vitesse}
\title{Vitesse}
\format{
## `vitesse`
Un vecteur de longueur 20.
\describe{
  \item{vitesse}{Mesure d'un score de vitesse en seconde}
}
}
\source{
Généré à l'aide du package {agk}.
}
\usage{
vitesse
}
\description{
Variable continue ayant 20 participants.
}
\keyword{datasets}
